clean ogv1
