+++ includes/utils.h
@@ -1,4 +1,3 @@
-
 #ifndef UTILS_H
 # define UTILS_H
 
@@ -10,6 +9,33 @@
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
+
+/* linux用 */
+#if defined(__linux__)
+enum	e_key_code
+{
+	KEY_W		= 119,
+	KEY_S		= 115,
+	KEY_A		= 97,
+	KEY_D		= 100,
+	/* KEY_ESC		= 27, */
+	KEY_ESC		= 65307,
+	KEY_LEFT	= 65361,
+	KEY_RIGHT	= 65363
+};
+#else
+/* mac用 */
+enum	e_key_code
+{
+	KEY_W		= 13,
+	KEY_S		= 1,
+	KEY_A		= 0,
+	KEY_D		= 2,
+	KEY_ESC		= 53,
+	KEY_LEFT	= 123,
+	KEY_RIGHT	= 124
+};
+#endif
 
 #define SYSERR	-1
 
@@ -28,11 +54,8 @@ enum	e_color
 
 typedef struct s_pos
 {
-	// int	x;
-	// int y;
 	double	x;
 	double	y;
 } t_pos;
-
 
 #endif
