+++ srcs/ogv1.c
@@ -1,6 +1,5 @@
 #include "ogv1.h"
 #include "calc.h"
-#include "config.h"
 #include "parse.h"
 #include "draw.h"
 
@@ -37,51 +36,6 @@ int	handle_keypress(int key, t_mlx *mlx)
 	render(mlx);
 	return (0);
 }
-	// double next_angle;
-	// t_pos res;
-		// mlx->player.pos.x -= MOVE_STEP;
-		// mlx->player.pos.x += MOVE_STEP * cos_wrap(fix_angle(mlx->player.angle + 90));
-		// mlx->player.pos.y -= MOVE_STEP * sin_wrap(fix_angle(mlx->player.angle + 90));
-		// res.x = mlx->player.pos.x;
-		// res.y = mlx->player.pos.y;
-		// res = fix_move(mlx, res, fix_angle(mlx->player.angle + 90.0), MOVE_STEP);
-		// mlx->player.pos.x = res.x;
-		// mlx->player.pos.y = res.y;
-		// mlx->player.pos.x += MOVE_STEP;
-		// mlx->player.pos.x += MOVE_STEP * cos_wrap(fix_angle(mlx->player.angle - 90));
-		// mlx->player.pos.y -= MOVE_STEP * sin_wrap(fix_angle(mlx->player.angle - 90));
-		// res.x = mlx->player.pos.x;
-		// res.y = mlx->player.pos.y;
-		// res = fix_move(mlx, res, fix_angle(mlx->player.angle - 90.0), MOVE_STEP);
-		// mlx->player.pos.x = res.x;
-		// mlx->player.pos.y = res.y;
-		// mlx->player.pos.y -= MOVE_STEP;
-		// mlx->player.pos.x += MOVE_STEP * cos_wrap(fix_angle(mlx->player.angle));
-		// mlx->player.pos.y -= MOVE_STEP * sin_wrap(fix_angle(mlx->player.angle));
-		// res.x = mlx->player.pos.x;
-		// res.y = mlx->player.pos.y;
-		// res = fix_move(mlx, res, mlx->player.angle, MOVE_STEP);
-		// mlx->player.pos.x = res.x;
-		// mlx->player.pos.y = res.y;
-		// mlx->player.pos.y += MOVE_STEP;
-		// mlx->player.pos.x += MOVE_STEP * cos_wrap(fix_angle(mlx->player.angle - 180));
-		// mlx->player.pos.y -= MOVE_STEP * sin_wrap(fix_angle(mlx->player.angle - 180));
-		// res.x = mlx->player.pos.x;
-		// res.y = mlx->player.pos.y;
-		// res = fix_move(mlx, res, fix_angle(mlx->player.angle - 180.0), MOVE_STEP);
-		// mlx->player.pos.x = res.x;
-		// mlx->player.pos.y = res.y;
-
-t_pos	convert_player_pos(t_loader loader)
-{
-	t_pos	pos;
-
-	pos.x = (loader.player_pos.x * MAP_PANEL_SIDE) + (MAP_PANEL_SIDE / 2);
-	pos.y = (loader.player_pos.y * MAP_PANEL_SIDE) + (MAP_PANEL_SIDE / 2);
-	// pos.x = 70;
-	// pos.y = 70;
-	return (pos);
-}
 
 void	initialize_render(t_mlx *mlx, t_loader loader)
 {
@@ -89,7 +43,7 @@ void	initialize_render(t_mlx *mlx, t_loader loader)
 	mlx->window = mlx_new_window(mlx->handle, WINDOW_WIDTH, WINDOW_HEIGHT, "cuv3d"); 
 	mlx->img.handle = mlx_new_image(mlx->handle, WINDOW_WIDTH, WINDOW_HEIGHT);
 	mlx->img.buffer = mlx_get_data_addr(mlx->img.handle, &(mlx->img.bits_per_pixel), &(mlx->img.line_size), &(mlx->img.endian));
-	mlx->player.pos = loader.player_pos;//convert_player_pos(loader)
+	mlx->player.pos = loader.player_pos;
 	mlx->player.angle = loader.player_angle;
 	mlx->map.data = ft_double_str_dup(loader.map_data);
 	mlx->map.y_count = loader.map_y_count;
@@ -116,51 +70,8 @@ int main(int argc, char* argv[])
 		return (1);
 	loader = parse(argv[1]);
 	initialize_render(&mlx, loader);
-	// printf("p.x=%lf, p.y=%lf, p.angle=%lf\n", mlx.player.pos.x, mlx.player.pos.y, mlx.player.angle);
-	//render
+	free_loader(loader);
 	render(&mlx);
-	//loop
 	mlx_hook(mlx.window, 02, 1L << 0, &handle_keypress, &mlx);
 	mlx_loop(mlx.handle);
 }
-	// mlx->player.side = 10;
-// void	print_mlx(t_mlx mlx)
-// {
-// 	prin
-// }
-
-// #include "string.h"
-
-
-	// mlx.handle = mlx_init(); //定数
-	// mlx.window = mlx_new_window(mlx.handle, WINDOW_WIDTH, WINDOW_HEIGHT, "cuv3d"); 
-	// mlx.player.pos.x = 251; //convert
-	// mlx.player.pos.y = 251;
-	// mlx.player.angle = 63.7;//63.7
-	// mlx.player.side = 10; //定数
-	// mlx.img.handle = mlx_new_image(mlx.handle, WINDOW_WIDTH, WINDOW_HEIGHT); //定数
-	// mlx.img.buffer = mlx_get_data_addr(mlx.img.handle, &mlx.img.bits_per_pixel, &mlx.img.line_size, &mlx.img.endian);
-	// mlx.map.x_count = MAP_X_COUNT;//loader.map.count <- まだ受け取ってない
-	// mlx.map.y_count = MAP_Y_COUNT; 
-	// mlx.map.panel_side = WINDOW_HEIGHT / MAP_Y_COUNT; //定数
-	// mlx.map.height = mlx.map.y_count * mlx.map.panel_side; //loader.map.count * 定数
-	// mlx.map.width =  mlx.map.x_count * mlx.map.panel_side; 
-	// mlx.map.data = (char **)malloc(sizeof(char *) * MAP_Y_COUNT); //loaderから直接
-	// mlx.map.data[0] = strdup("1111111111");
-	// mlx.map.data[1] = strdup("1010001001");
-	// mlx.map.data[2] = strdup("1010001001");
-	// mlx.map.data[3] = strdup("1010000001");
-	// mlx.map.data[4] = strdup("1000000001");
-	// mlx.map.data[5] = strdup("1000100001");
-	// mlx.map.data[6] = strdup("1000100001");
-	// mlx.map.data[7] = strdup("1000000001");
-	// mlx.map.data[8] = strdup("1000000101");
-	// mlx.map.data[9] = strdup("1111111111");
-	// mlx.textures.north.handle = mlx_xpm_file_to_image(mlx.handle, NORTH_TEX_PATH, &(mlx.textures.north.width), &(mlx.textures.north.height));
-	// mlx.textures.north.buffer = mlx_get_data_addr(mlx.textures.north.handle, &(mlx.textures.north.bits_per_pixel), &(mlx.textures.north.line_size), &(mlx.textures.north.endian));
-	// mlx.textures.south.handle = mlx_xpm_file_to_image(mlx.handle, SOUTH_TEX_PATH, &(mlx.textures.south.width), &(mlx.textures.south.height));
-	// mlx.textures.south.buffer = mlx_get_data_addr(mlx.textures.south.handle, &(mlx.textures.south.bits_per_pixel), &(mlx.textures.south.line_size), &(mlx.textures.south.endian));
-	// mlx.textures.west.handle = mlx_xpm_file_to_image(mlx.handle, WEST_TEX_PATH, &(mlx.textures.west.width), &(mlx.textures.west.height));
-	// mlx.textures.west.buffer = mlx_get_data_addr(mlx.textures.west.handle, &(mlx.textures.west.bits_per_pixel), &(mlx.textures.west.line_size), &(mlx.textures.west.endian));
-	// mlx.textures.east.handle = mlx_xpm_file_to_image(mlx.handle, EAST_TEX_PATH, &(mlx.textures.east.width), &(mlx.textures.east.height));
-	// mlx.textures.east.buffer = mlx_get_data_addr(mlx.textures.east.handle, &(mlx.textures.east.bits_per_pixel), &(mlx.textures.east.line_size), &(mlx.textures.east.endian));
